Modigied

+++ Src/ShopManagement/1.Core/ShopManagement.Core.RequestResponse/Products/Command/Update/UpdateProductCommand.cs
@@ -13,7 +13,7 @@ namespace ShopManagement.Core.RequestResponse.Products.Command.Update
         public string Picture { get; set; }
         public string PictureAlt { get; set; }
         public string PictureTitle { get; set; }
-        public long CategoryId { get; set; }
+        public int CategoryId { get; set; }
         public string Slug { get; set; }
         public string Keywords { get; set; }
         public string MetaDescription { get; set; }
